+++ database/migrations/2024_10_13_115602_create_role_user_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('role_user', function (Blueprint $table) {
+        if (!Schema::hasTable('role_user')) {
+            Schema::create('role_user', function (Blueprint $table) {
             $table->id(); // Add the primary key column
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('role_id')->constrained()->onDelete('cascade');
             $table->unique(['user_id', 'role_id']);
-        });
+            });
+        }
     }
 
     public function down()
